Create ellama-help-with-edit-op function and ellama hydra binding.
This function gets the ellama provider to tell the user how to
conveniently perform a desired task in emacs (such as an arbitrary
command that would be useful to perform conveniently, eg: list open
child processes, close child processes, indent multiple lines of text
at once, etc. If no convenient key sequence exists to perform the
task, the ellama provider will define try to define an elisp function
to perform it, which can manually be saved in the user's init file and
given a convenient key binding.